zidio task optimizer 2/requirements.txt
flask
flask-cors
nltk
opencv-python
tensorflow-macos
SpeechRecognition
numpy
dlib
psycopg2-binary
